+++ src/main/java/io/github/sakurawald/module/initializer/color/anvil/config/anvil/ColorAnvilConfigModel.java
@@ -1,0 +1,5 @@
+package io.github.sakurawald.module.initializer.color.anvil.config.anvil;
+
+public class ColorAnvilConfigModel {
+    public boolean requires_corresponding_permission_to_use_style_tag = false;
+}