+++ src/main/java/io/github/sakurawald/module/initializer/color/sign/ColorSignInitializer.java
@@ -3,17 +3,21 @@ package io.github.sakurawald.module.initializer.color.sign;
 import io.github.sakurawald.core.auxiliary.LogUtil;
 import io.github.sakurawald.core.auxiliary.minecraft.UuidHelper;
 import io.github.sakurawald.core.config.handler.abst.BaseConfigurationHandler;
+import io.github.sakurawald.core.config.handler.impl.ObjectConfigurationHandler;
 import io.github.sakurawald.core.manager.Managers;
 import io.github.sakurawald.core.structure.SpatialBlock;
 import io.github.sakurawald.module.initializer.ModuleInitializer;
+import io.github.sakurawald.module.initializer.color.sign.config.model.ColorSignConfigModel;
 import io.github.sakurawald.module.initializer.color.sign.structure.SignCache;
 import org.jetbrains.annotations.Nullable;
 
 import java.io.IOException;
 
-public class SignInitializer extends ModuleInitializer {
+public class ColorSignInitializer extends ModuleInitializer {
 
     private static final String ATTACHMENT_SUBJECT = "color-sign-cache";
+
+    public static final BaseConfigurationHandler<ColorSignConfigModel> config = new ObjectConfigurationHandler<>(BaseConfigurationHandler.CONFIG_JSON, ColorSignConfigModel.class);
 
     public static @Nullable SignCache readSignCache(SpatialBlock spatialBlock) {
         String uuid = UuidHelper.getAttachedUuid(spatialBlock);