+++ src/main/java/io/github/sakurawald/module/initializer/color/sign/config/model/ColorSignConfigModel.java
@@ -1,0 +1,5 @@
+package io.github.sakurawald.module.initializer.color.sign.config.model;
+
+public class ColorSignConfigModel {
+    public boolean requires_corresponding_permission_to_use_style_tag = false;
+}